pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "alphazoo"
version = "0.1.0"
description = "Standalone AlphaZero implementation with PettingZoo compatibility"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Guilherme"}
]
keywords = ["reinforcement-learning", "alphazero", "mcts", "pettingzoo", "self-play"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "pettingzoo>=1.24.0",
    "gymnasium>=0.29.0",
    "ray>=2.7.0",
    "scipy>=1.11.0",
    "ruamel.yaml>=0.17.0",
    "psutil>=5.9.0",
    "termcolor>=2.3.0",
    "progress>=1.6",
    "more_itertools>=9.0.0",
    "cityhash>=0.2.3",
    "metrohash>=0.0.6",
    "bitstring>=3.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
]
plotting = [
    "matplotlib>=3.7.0",
]

[project.urls]
Homepage = "https://github.com/guilherme439/alphazoo"
Documentation = "https://github.com/guilherme439/alphazoo#readme"
Repository = "https://github.com/guilherme439/alphazoo"
"Bug Tracker" = "https://github.com/guilherme439/alphazoo/issues"

[tool.setuptools.package-dir]
alphazoo = "main"

[tool.setuptools]
packages = [
    "alphazoo",
    "alphazoo.configs",
    "alphazoo.search",
    "alphazoo.training",
    "alphazoo.testing",
    "alphazoo.testing.agents",
    "alphazoo.testing.agents.generic",
    "alphazoo.utils",
    "alphazoo.utils.caches",
    "alphazoo.utils.functions",
    "alphazoo.utils.progress_bars",
    "alphazoo.wrappers",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
